requirements.txt
aniso8601==7.0.0
arrow==0.17.0
asgiref==3.3.0
boto3==1.16.10
boto3-stubs==1.16.10.0
botocore==1.19.10
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.5
dateparser==1.0.0
Django==3.1.5
django-cleanup==5.1.0
django-cors-headers==3.6.0
django-filter==2.4.0
django-graphql-auth==0.3.15
django-graphql-jwt==0.3.0
django-graphql-playground==1.0
django-storages==1.10.1
django-webpush==0.3.3
djangorestframework==3.12.2
djangorestframework-camel-case==1.2.0
docutils==0.16
docxtpl==0.11.3
graphene==2.1.8
graphene-django==2.15.0
graphene-django-cud==0.7.0
graphene-django-optimizer==0.8.0
graphene-file-upload==1.2.2
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
http-ece==1.1.0
idna==2.10
Jinja2==2.11.3
jmespath==0.10.0
lxml==4.6.2
MarkupSafe==1.1.1
moment==0.12.1
Pillow==8.1.0
promise==2.3
psycopg2-binary==2.8.6
py-vapid==1.7.1
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.5
pywebpush==1.9.4
regex==2020.11.13
requests==2.25.1
Rx==1.6.1
s3transfer==0.3.3
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
times==0.7
typing-extensions==3.7.4.3
tzlocal==2.1
Unidecode==1.1.1
urllib3==1.25.11
whitenoise==5.2.0
